+++ src/Components/Tabs/Tab.tsx
@@ -3,7 +3,7 @@ import React, {forwardRef, MouseEvent} from 'react'
 import classnames from 'classnames'
 
 // Types
-import {StandardFunctionProps} from '../../Types'
+import {StandardFunctionProps, RenderLinkElement} from '../../Types'
 
 export interface TabProps extends StandardFunctionProps {
   /** Renders the tab highlighted */
@@ -15,9 +15,11 @@ export interface TabProps extends StandardFunctionProps {
   /** Icon to appear left of the text label */
   icon?: JSX.Element
   /** Function to call when tab is clicked, id of tab is passed in */
-  onClick: (id?: string) => void
+  onClick?: (id?: string) => void
   /** If a function is passed in a dismiss button is rendered in the right of the tab */
   onDismiss?: (id?: string) => void
+  /** Optional link element. Will override onClick prop */
+  linkElement?: RenderLinkElement
 }
 
 export type TabRef = HTMLDivElement
@@ -29,18 +31,21 @@ export const Tab = forwardRef<TabRef, TabProps>(
       icon,
       text,
       style,
+      testID = 'tabs--tab',
       active,
       onClick,
       className,
       onDismiss,
-      testID = 'tabs--tab',
+      linkElement,
     },
     ref
   ) => {
     const handleClick = (e: MouseEvent<HTMLDivElement>): void => {
       e.stopPropagation()
 
-      onClick(id)
+      if (onClick) {
+        onClick(id)
+      }
     }
 
     const handleDismissClick = (e: MouseEvent<HTMLButtonElement>): void => {
@@ -57,15 +62,8 @@ export const Tab = forwardRef<TabRef, TabProps>(
       [`${className}`]: className,
     })
 
-    return (
-      <div
-        ref={ref}
-        className={tabClass}
-        data-testid={testID}
-        id={id}
-        style={style}
-        onClick={handleClick}
-      >
+    const tabContents = (
+      <>
         {icon}
         <span className="cf-tabs--tab-label">{text}</span>
         {onDismiss && (
@@ -77,6 +75,28 @@ export const Tab = forwardRef<TabRef, TabProps>(
             <div className="cf-tabs--tab-dismiss-circle" />
           </button>
         )}
+        <div className="cf-tabs--state-indicator" />
+      </>
+    )
+
+    if (linkElement) {
+      return React.cloneElement(
+        linkElement(tabClass),
+        {'data-testid': testID},
+        tabContents
+      )
+    }
+
+    return (
+      <div
+        ref={ref}
+        className={tabClass}
+        data-testid={testID}
+        id={id}
+        style={style}
+        onClick={handleClick}
+      >
+        {tabContents}
       </div>
     )
   }