+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#464](https://github.com/influxdata/clockface/pull/464): Allow optional `shortLabel` on `TreeNavItem` strictly for display when `TreeNav` is collapsed
 - [#464](https://github.com/influxdata/clockface/pull/464): Make `TreeNavSubMenu` visible so long as the entire `TreeNav` is visible
 - [#464](https://github.com/influxdata/clockface/pull/464): Allow `TreeNav` to be opened and closed on small screens independently of the `expanded` prop
+- [#464](https://github.com/influxdata/clockface/pull/464): Allow banner element to optionally remain visible when `TreeNav` is collapsed
 
 #### 2.0.2