+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 2.0.3 (Unreleased)
 
 - [#466](https://github.com/influxdata/clockface/pull/466): Fix `OverlayHeading` overflow issue
+- [#466](https://github.com/influxdata/clockface/pull/466): Enable optional text wrapping in `OverlayHeading`
 - [#466](https://github.com/influxdata/clockface/pull/466): Fix appearance of dismiss buttons in `NotificationDialog`
 - [#466](https://github.com/influxdata/clockface/pull/466): Fix appearance of toggle in `VisibilityInput`
 - [#466](https://github.com/influxdata/clockface/pull/466): Ensure `Heading` component does not remove margins specified in the funnel page heading styles