+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 2.0.3 (Unreleased)
 
+- [#469](https://github.com/influxdata/clockface/pull/469): Add optional `linkElement` prop to `Tab`
 - [#466](https://github.com/influxdata/clockface/pull/466): Fix `OverlayHeading` overflow issue
 - [#466](https://github.com/influxdata/clockface/pull/466): Enable optional text wrapping in `OverlayHeading`
 - [#466](https://github.com/influxdata/clockface/pull/466): Fix appearance of dismiss buttons in `NotificationDialog`